refactor: using src/ tree format
