--- clock/__init__.py
@@ -1,3 +1,0 @@
-from .clock import Clock
-
-__all__ = ["Clock"]
