--- clock/exceptions.py
@@ -1,2 +1,0 @@
-class IntervalNotSupported(Exception):
-    """Raised when the supplied interval is not supported/allowed"""
