--- consts/__init__.py
@@ -1,7 +1,0 @@
-from .consts import *
-
-__all__ = [
-    "TIME_FMT_DAY",
-    "TIME_FMT_FULL",
-    "TIME_FMT_MINS",
-]
