--- models/exceptions.py
@@ -1,10 +1,0 @@
-class IdenticalSMASCantCrossError(Exception):
-    pass
-
-
-class WTF(Exception):
-    """SHOULD NEVER be raised"""
-
-
-class EmptyPriceActionError(Exception):
-    pass
