+++ src/backtests/config.py
@@ -1,3 +1,5 @@
+"""This file holds default configurations and constant variables to be used globally"""
+
 TIME_FMT_DAY = "%Y-%m-%d"
 TIME_FMT_FULL = "%Y-%m-%dT%H:%M:%S"
 TIME_FMT_MINS = "%Y-%m-%dT%H:%M:%S"
