+++ src/backtests/core/__init__.py
@@ -1,2 +1,4 @@
 from .price_action import PriceAction
 from .price_bar import Bar
+from .clock import Clock
+from .client import YClient
