+++ src/backtests/strategies/method.py
