+++ src/backtests/trader/README.md
