--- utils/__init__.py
@@ -1,3 +1,0 @@
-from .funcs import parse_interval, td_to_str, discard_datetime_by_interval
-
-__all__ = ["parse_interval", "td_to_str", "discard_datetime_by_interval"]
