--- utils/exceptions.py
@@ -1,2 +1,0 @@
-class InvalidIntervalFormat(Exception):
-    """Raised when a str representation of a time interval is invalid"""
